@@ -1,19 +1,32 @@
-import path from 'path';
+import path from 'node:path';
+import process from 'node:process';
+import {fileURLToPath} from 'node:url';
 import test from 'ava';
-import m from '.';
+import binCheck from './index.js';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const __filename = fileURLToPath(import.meta.url);
 
 const bin = {
 	darwin: path.join(__dirname, 'fixtures/optipng-osx'),
 	linux: path.join(__dirname, 'fixtures/optipng-linux'),
-	win32: path.join(__dirname, 'fixtures/optipng-win32.exe')
+	win32: path.join(__dirname, 'fixtures/optipng-win32.exe'),
 };
 
 test('async', async t => {
-	t.true(await m(bin[process.platform]));
-	await t.throws(m(__filename), `Couldn't execute the \`${__filename}\` binary. Make sure it has the right permissions.`);
+	t.true(await binCheck(bin[process.platform]));
+	await t.throwsAsync(
+		binCheck('node', [__filename]),
+		undefined,
+		`Couldn't execute the "${__filename}" binary. Make sure it has the right permissions.`,
+	);
 });
 
 test('sync', t => {
-	t.true(m.sync(bin[process.platform]));
-	t.throws(m.sync.bind(null, __filename), `Couldn't execute the \`${__filename}\` binary. Make sure it has the right permissions.`);
+	t.true(binCheck.sync(bin[process.platform]));
+	t.throws(
+		binCheck.sync.bind(null, 'node', [__filename]),
+		undefined,
+		`Couldn't execute the "${__filename}" binary. Make sure it has the right permissions.`,
+	);
 });
